hme_tvqa/pytorch_c3d_extraction/requirements.txt
backports-abc==0.5
backports.functools-lru-cache==1.6.1
backports.shutil-get-terminal-size==1.0.0
bootstrap.pytorch==0.0.13
boto3==1.10.48
botocore==1.13.48
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
colorlog==4.0.2
cycler==0.10.0
decorator==4.4.1
docutils==0.15.2
enum34==1.1.6
funcsigs==1.0.2
future==0.18.2
futures==3.3.0
h5py==2.10.0
idna==2.8
ipdb==0.12.3
ipython==5.8.0
ipython-genutils==0.2.0
jmespath==0.9.4
joblib==0.14.1
jsonpatch==1.24
jsonpointer==2.0
kiwisolver==1.1.0
matplotlib==2.2.4
mock==3.0.5
munch==2.5.0
networkx==2.2
numexpr==2.7.0
numpy==1.16.5
opencv-python==4.1.2.30
pandas==0.24.2
pathlib2==2.3.5
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.2
plotly==4.4.1
pretrainedmodels==0.7.4
prompt-toolkit==1.0.18
protobuf==3.11.2
ptyprocess==0.6.0
Pygments==2.5.2
pyparsing==2.4.6
python-dateutil==2.8.1
pytorch-pretrained-bert==0.6.2
pytz==2019.3
PyWavelets==1.0.3
PyYAML==5.3
pyzmq==18.1.1
regex==2020.1.8
requests==2.22.0
retrying==1.3.3
s3transfer==0.2.1
sacremoses==0.0.38
scandir==1.10.0
scikit-image==0.14.5
scipy==1.2.2
seaborn==0.9.0
sentencepiece==0.1.85
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.13.0
skipthoughts==0.0.1
subprocess32==3.5.4
tables==3.5.2
tabulate==0.8.6
tensorboardX==2.0
torch==1.3.1
torchfile==0.1.0
torchvision==0.4.2
tornado==5.1.1
tqdm==4.41.1
traitlets==4.3.3
urllib3==1.25.7
visdom==0.1.8.9
wcwidth==0.1.8
websocket-client==0.57.0
